+++ Common/src/main/java/com/tristankechlo/livingthings/LivingThings.java
@@ -1,16 +1,24 @@
 package com.tristankechlo.livingthings;
 
 import com.tristankechlo.livingthings.entity.*;
+import com.tristankechlo.livingthings.entity.projectile.ThrownOstrichEgg;
 import com.tristankechlo.livingthings.init.ModBlocks;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 import com.tristankechlo.livingthings.init.ModItems;
 import com.tristankechlo.livingthings.init.ModSounds;
 import com.tristankechlo.livingthings.mixin.SpawnPlacementsInvoker;
+import net.minecraft.Util;
+import net.minecraft.core.Position;
+import net.minecraft.core.dispenser.AbstractProjectileDispenseBehavior;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.SpawnPlacements.Type;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.monster.Monster;
+import net.minecraft.world.entity.projectile.Projectile;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.DispenserBlock;
 import net.minecraft.world.level.levelgen.Heightmap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -77,6 +85,14 @@ public final class LivingThings {
         SpawnPlacementsInvoker.register(ModEntityTypes.SEAHORSE.get(), Type.IN_WATER, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, SeahorseEntity::checkSeahorseSpawnRules);
         SpawnPlacementsInvoker.register(ModEntityTypes.BABY_ENDER_DRAGON.get(), Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, BabyEnderDragonEntity::checkBabyEnderDragonSpawnRules);
         SpawnPlacementsInvoker.register(ModEntityTypes.PEACOCK.get(), Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, PeacockEntity::checkPeacockSpawnRules);
+    }
+
+    public static void registerDispenserBehavior() {
+        DispenserBlock.registerBehavior(ModItems.OSTRICH_EGG.get(), new AbstractProjectileDispenseBehavior() {
+            protected Projectile getProjectile(Level world, Position pos, ItemStack stack) {
+                return Util.make(new ThrownOstrichEgg(world, pos.x(), pos.y(), pos.z()), (egg) -> egg.setItem(stack));
+            }
+        });
     }
 
 }