+++ Common/src/main/java/com/tristankechlo/livingthings/item/OstrichEggItem.java
@@ -1,0 +1,39 @@
+package com.tristankechlo.livingthings.item;
+
+import com.tristankechlo.livingthings.entity.projectile.ThrownOstrichEgg;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.stats.Stats;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+
+public class OstrichEggItem extends Item {
+
+    public OstrichEggItem(Properties props) {
+        super(props);
+    }
+
+    @Override
+    public InteractionResultHolder<ItemStack> use(Level world, Player player, InteractionHand hand) {
+        ItemStack stack = player.getItemInHand(hand);
+        world.playSound(null, player.getX(), player.getY(), player.getZ(), SoundEvents.EGG_THROW, SoundSource.PLAYERS, 0.5F, 0.4F / (world.getRandom().nextFloat() * 0.4F + 0.8F));
+        if (!world.isClientSide) {
+            ThrownOstrichEgg thrownEgg = new ThrownOstrichEgg(world, player);
+            thrownEgg.setItem(stack);
+            thrownEgg.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, 1.5F, 1.0F);
+            world.addFreshEntity(thrownEgg);
+        }
+
+        player.awardStat(Stats.ITEM_USED.get(this));
+        if (!player.getAbilities().instabuild) {
+            stack.shrink(1);
+        }
+
+        return InteractionResultHolder.sidedSuccess(stack, world.isClientSide());
+    }
+
+}